Move canteenbot-core into own monorepo module
