--- core/test.json
@@ -1,7 +1,0 @@
-{
-  "httpMethod": "GET",
-  "queryStringParameters": {
-    "message_type": "INGREDIENT",
-    "message_param": "fish"
-  }
-}
